Exception package modified

+++ src/main/java/com/alkemy/ong/service/impl/MemberServiceImpl.java
@@ -2,25 +2,24 @@ package com.alkemy.ong.service.impl;
 
 import com.alkemy.ong.dto.MemberDTO;
 import com.alkemy.ong.dto.response.PostMembersDTO;
+import com.alkemy.ong.exception.EntityNotFoundException;
 import com.alkemy.ong.exception.NullListException;
 import com.alkemy.ong.mapper.MemberMapper;
 import com.alkemy.ong.model.Member;
 import com.alkemy.ong.repository.MemberRepository;
 import com.alkemy.ong.service.IMemberService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
 import org.springframework.stereotype.Service;
 
-import lombok.RequiredArgsConstructor;
-
-import javax.persistence.EntityNotFoundException;
 import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Locale;
 
 @RequiredArgsConstructor
 @Service
-public class MemberServiceImpl implements IMemberService{
+public class MemberServiceImpl implements IMemberService {
     private final MemberRepository memberRepository;
 
     @Autowired
@@ -30,17 +29,17 @@ public class MemberServiceImpl implements IMemberService{
 
 
     @Override
-    public PostMembersDTO createMember(MemberDTO memberDTO){
+    public PostMembersDTO createMember(MemberDTO memberDTO) {
         MemberMapper member = new MemberMapper();
 
         Member memberSaved = memberRepository.save(member.mapperMember(memberDTO));
 
         PostMembersDTO response = new PostMembersDTO();
         response.setId(memberSaved.getId());
-        response.setUrl("/members/"+memberSaved.getId());
+        response.setUrl("/members/" + memberSaved.getId());
 
         return response;
-        
+
     }
 
     @Override
@@ -55,9 +54,7 @@ public class MemberServiceImpl implements IMemberService{
     @Override
     @Transactional
     public void delete(Long id) {
-        memberRepository.findById(id)
-                .orElseThrow(()->
-                        new EntityNotFoundException(message.getMessage("data.not.found", null,Locale.US)));
+        if (!memberRepository.existsById(id)) throw new EntityNotFoundException("Members", "id", id);
         memberRepository.deleteById(id);
     }
